+++ src/models/Discount.php
@@ -280,11 +280,30 @@ class Discount extends Model
     }
 
     /**
+     * @return bool
+     * @since 4.3.0
+     */
+    public function hasOrderCondition(): bool
+    {
+        if ($this->_orderCondition === null) {
+            return false;
+        }
+
+        return empty($this->getOrderCondition()->getConditionRules());
+    }
+
+    /**
      * @param ElementConditionInterface|string|array $condition
      * @return void
+     * @throws InvalidConfigException
      */
     public function setOrderCondition(ElementConditionInterface|string|array $condition): void
     {
+        if (empty($condition)) {
+            $this->_orderCondition = null;
+            return;
+        }
+
         if (is_string($condition)) {
             $condition = Json::decodeIfJson($condition);
         }
@@ -312,11 +331,30 @@ class Discount extends Model
     }
 
     /**
-     * @param ElementConditionInterface|string|array $condition
-     * @return void
+     * @return bool
+     * @since 4.3.0
      */
-    public function setCustomerCondition(ElementConditionInterface|string|array $condition): void
+    public function hasCustomerCondition(): bool
     {
+        if ($this->_orderCondition === null) {
+            return false;
+        }
+
+        return empty($this->getCustomerCondition()->getConditionRules());
+    }
+
+    /**
+     * @param ElementConditionInterface|string $condition
+     * @return void
+     * @throws InvalidConfigException
+     */
+    public function setCustomerCondition(ElementConditionInterface|string $condition): void
+    {
+        if (empty($condition)) {
+            $this->_customerCondition = null;
+            return;
+        }
+
         if (is_string($condition)) {
             $condition = Json::decodeIfJson($condition);
         }
@@ -345,11 +383,30 @@ class Discount extends Model
     }
 
     /**
+     * @return bool
+     * @since 4.3.0
+     */
+    public function hasShippingAddressCondition(): bool
+    {
+        if ($this->_orderCondition === null) {
+            return false;
+        }
+
+        return empty($this->getShippingAddressCondition()->getConditionRules());
+    }
+
+    /**
      * @param ElementConditionInterface|string|array $condition
      * @return void
+     * @throws InvalidConfigException
      */
     public function setShippingAddressCondition(ElementConditionInterface|string|array $condition): void
     {
+        if (empty($condition)) {
+            $this->_shippingAddressCondition = null;
+            return;
+        }
+
         if (is_string($condition)) {
             $condition = Json::decodeIfJson($condition);
         }
@@ -378,11 +435,30 @@ class Discount extends Model
     }
 
     /**
+     * @return bool
+     * @since 4.3.0
+     */
+    public function hasBillingAddressCondition(): bool
+    {
+        if ($this->_orderCondition === null) {
+            return false;
+        }
+
+        return empty($this->getBillingAddressCondition()->getConditionRules());
+    }
+
+    /**
      * @param ElementConditionInterface|string|array $condition
      * @return void
+     * @throws InvalidConfigException
      */
     public function setBillingAddressCondition(ElementConditionInterface|string|array $condition): void
     {
+        if (empty($condition)) {
+            $this->_billingAddressCondition = null;
+            return;
+        }
+
         if (is_string($condition)) {
             $condition = Json::decodeIfJson($condition);
         }
@@ -552,6 +628,45 @@ class Discount extends Model
                     }
                 },
             ],
+            [[
+                'id',
+                'allPurchasables',
+                'allCategories',
+                'purchasableIds',
+                'categoryIds',
+                'name',
+                'description',
+                'couponFormat',
+                'orderCondition',
+                'customerCondition',
+                'billingAddressCondition',
+                'shippingAddressCondition',
+                'perUserLimit',
+                'perEmailLimit',
+                'totalDiscountUseLimit',
+                'totalDiscountUses',
+                'dateFrom',
+                'dateTo',
+                'purchaseTotal',
+                'orderConditionFormula',
+                'purchaseQty',
+                'maxPurchaseQty',
+                'baseDiscount',
+                'baseDiscountType',
+                'perItemDiscount',
+                'percentDiscount',
+                'percentageOffSubject',
+                'excludeOnSale',
+                'hasFreeShippingForMatchingItems',
+                'hasFreeShippingForOrder',
+                'categoryRelationshipType',
+                'enabled',
+                'stopProcessing',
+                'dateCreated',
+                'dateUpdated',
+                'ignoreSales',
+                'appliedTo',
+            ], 'safe'],
         ];
     }